Refactor commands an providers registration
